+++ doc/gb-model/release_notes.rst
@@ -12,6 +12,7 @@ Release Notes
 Upcoming Release
 ================
 
+* Add demands to pypsa Network (#102, #70).
 * Limit GB model to ``clustered`` clusters.
 * Add EV to pypsa Network (#114)
 * Tablulated regional unmanaged EV charging demand data (#112).